pyproject.toml
[project]
name = "python-mcp"
version = "0.1.0"
description = "Python Voice Backend Server with MCP"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.5",
    "loguru>=0.7.2",
    "openai>=1.35.0",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.10.0,<1.14.0",
    "soundfile>=0.12.1",
    "librosa>=0.10.0,<0.11.0",
    "llvmlite>=0.40.0",
    "numba>=0.57.0",
    "python-dateutil>=2.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[dependency-groups]
dev = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "opuslib-next>=1.1.0",
]
